+++ rails/changelog.md
@@ -5,6 +5,7 @@
 ### Enhancements
 - Added multiple and tag Select variants
 - Added a placeholder attribute for the Select
+- Added the component version in the tooltip of each component
 - Adjustments to the workspace display and overlays
 - Project and workspace names are now sorted alphabetically
 - Enabled IE 10 support for Kaiju Previews